+++ noodles-bam/src/bai/writer.rs
@@ -2,8 +2,9 @@ use std::io::{self, Write};
 
 use byteorder::{LittleEndian, WriteBytesExt};
 use noodles_csi::{
+    binning_index::ReferenceSequenceExt,
     index::reference_sequence::{bin::Chunk, Metadata},
-    BinningIndex, BinningIndexReferenceSequence,
+    BinningIndex,
 };
 
 use super::{