+++ noodles-bcf/src/reader.rs
@@ -16,7 +16,7 @@ use std::{
 use byteorder::{LittleEndian, ReadBytesExt};
 use noodles_bgzf as bgzf;
 use noodles_core::{region::Interval, Region};
-use noodles_csi::{BinningIndex, BinningIndexReferenceSequence};
+use noodles_csi::{binning_index::ReferenceSequenceExt, BinningIndex};
 use noodles_vcf::header::Contigs;
 
 use super::Record;
@@ -254,15 +254,15 @@ where
     /// }
     /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
-    pub fn query<I, IRS>(
+    pub fn query<I, RS>(
         &mut self,
         contigs: &Contigs,
         index: &I,
         region: &Region,
     ) -> io::Result<Query<'_, R>>
     where
-        I: BinningIndex<IRS>,
-        IRS: BinningIndexReferenceSequence,
+        I: BinningIndex<RS>,
+        RS: ReferenceSequenceExt,
     {
         let (reference_sequence_id, interval) = resolve_region(contigs, region)?;
         let chunks = index.query(reference_sequence_id, interval)?;