+++ noodles-csi/src/lib.rs
@@ -10,12 +10,13 @@ pub mod index;
 mod reader;
 mod writer;
 
-pub use self::{
-    binning_index::{BinningIndex, BinningIndexReferenceSequence},
-    index::Index,
-    reader::Reader,
-    writer::Writer,
-};
+pub use self::{binning_index::BinningIndex, index::Index, reader::Reader, writer::Writer};
+
+#[deprecated(
+    since = "0.4.0",
+    note = "Use `csi::binning_index::ReferenceSequenceExt` instead."
+)]
+pub use binning_index::ReferenceSequenceExt as BinningIndexReferenceSequence;
 
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};